refactor: TransactionManager
