+++ src/main/java/com/lxp/Main.java
@@ -1,13 +1,13 @@
 package com.lxp;
 
-import com.lxp.component.ApplicationContext;
-import com.lxp.component.DatabaseInitializer;
+import com.lxp.config.JDBCConnection;
+import com.lxp.config.TransactionManager;
 import com.lxp.util.CLIRouter;
 
 public class Main {
     public static void main(String[] args) {
-        DatabaseInitializer.initialize();
-        CLIRouter router = ApplicationContext.getRouter();
+        TransactionManager.DatabaseInitializer.initialize();
+        CLIRouter router = JDBCConnection.ApplicationContext.getRouter();
         router.start();
     }
 }
